refactor(OrganizeEventUseCase test): create mocks
